% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crep_cies_pb.R
\name{crep_cies_pb}
\alias{crep_cies_pb}
\title{crep_cies_pb}
\usage{
crep_cies_pb(pi, pc, N = 0)
}
\arguments{
\item{pi}{The probability of a given result in the intervention group.}

\item{pc}{The probability of a given result in the control group.}

\item{N}{The total sample size.}
}
\value{
Uncorr. CIES: the CIES uncorrected for sample size
Corr. CIES: The CIES corrected for sample size
}
\description{
Calculates the N-corrected Cox index effect size when given the
probability of a given occurrence in two groups. The first two arguments
are the probability in the intervention group and control group,
respectively. Both should be decimals with a value < 1.00.

The third argument (optional) is the total N for the dataset. If this
argument is supplied, the correction for sample size in CIES (omega) will
be applied. If it is not supplied, omega will not be applied. Omega is
missing from the WWC handbook, but discussed in their online training
for baseline equivalence.
}
\examples{
crep_cies_pb(0.5, 0.7)
crep_cies_pb(0.6, 0.4, N = 10)
}
