% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crep_desc.R, R/crep_descriptives.R
\name{crep_desc}
\alias{crep_desc}
\title{crep_desc}
\usage{
crep_desc(df)

crep_desc(df)
}
\arguments{
\item{df}{A dataframe to run descriptives on.}
}
\value{
\code{crep_desc()} returns a list of two tables. The first is the results for
any factors. The second is the results for any numeric variables.

\code{crep_desc()} returns a list of two tables. The first is the results for
any factors. The second is the results for any numeric variables.
}
\description{
Calculates descriptives on all variables in the
supplied dataframe.

Calculates descriptives on all variables in the
supplied dataframe.
}
\details{
Variables as separated as categorical or numeric based on their datatype.
If a numeric variable represents nominal categories, it should be encoded
as a factor with \code{as.factor()} before using \code{crep_desc()}.

For each categorical variable, the number of levels and number of NAs is
reported. For each numeric variable, range, median, mean, and standard
deviation are supplied. Additionally, skewness and kurtosis are calculated,
and the robust Jarque-Bera test of normality is run. NA counts are
also reported.

Variables as separated as categorical or numeric based on their datatype.
If a numeric variable represents nominal categories, it should be encoded
as a factor with \code{as.factor()} before using \code{crep_desc()}.

For each categorical variable, the number of levels and number of NAs is
reported. For each numeric variable, range, median, mean, and standard
deviation are supplied. Additionally, skewness and kurtosis are calculated,
and the robust Jarque-Bera test of normality is run. NA counts are
also reported.
}
\examples{
x <- runif(10)
y <- runif(10)
df <- data.frame(x, y)
crep_desc(df)
x <- runif(10)
y <- runif(10)
df <- data.frame(x, y)
crep_desc(df)
}
