% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crep_ii.R
\name{crep_ii}
\alias{crep_ii}
\title{crep_ii}
\usage{
crep_ii(es)
}
\arguments{
\item{es}{Effect size}
}
\value{
percentile rank, improvement index
}
\description{
Calculates the percentile rank and improvement index
according to WWC methods. The argument is g or a g-equivalent effect
size such as the CIES. Other effect sizes may be used if translated
to g first.
}
\examples{

CREP::crep_ii(0.25)
}
