% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crep_scale.R
\name{crep_scale}
\alias{crep_scale}
\title{crep_scale}
\usage{
crep_scale(df, vars, type = c("standardize", "center"))
}
\arguments{
\item{df}{A dataframe}

\item{vars}{A string list with the names of the columns to be scaled}

\item{type}{The default, \code{"standardize"}, calculates z-scores for each
variable. \code{"center"} will subtract the mean of a variable from each of
its values.}
}
\value{
crep_scale returns a new dataframe with the standardized variables
attached.
}
\description{
When given a dataframe and a list of relevant columns, will
create new variables that either standardize (z-scores) or center
the supplied columns.
}
\examples{
crep_scale(crep_som, c("rating"), "center")

crep_scale(crep_somwide, c("rater1", "rater2"), "standardize")
}
