% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crep_som.R
\docType{data}
\name{crep_som}
\alias{crep_som}
\title{crep_som}
\format{
A data frame with 52 rows and 3 variables
}
\usage{
data(crep_som)
}
\description{
This is false data for the SOM, i.e., not reflective
of actual observations. Its purpose is to serve as a stand-in for
that data when writing or testing code.

This data is long. For wide data, see crep_somwide.
}
\keyword{datasets}
