% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crep_somwide.R
\docType{data}
\name{crep_somwide}
\alias{crep_somwide}
\title{crep_somwide}
\format{
A data frame with 26 rows and 2 variables
}
\usage{
data(crep_somwide)
}
\description{
This is false data for the SOM, i.e., not reflective
of actual observations. Its purpose is to serve as a stand-in for
that data when writing or testing code.

This data is wide. For long data, see crep_som.
}
\keyword{datasets}
